Selected enum added to CrisPaintView header CTypes header erased
