--- CTypes.h
@@ -1,14 +1,0 @@
-#pragma once
-
-enum Primitive_Selection
-{
-	NOTHING_SELECTED,
-	LINE_SELECTED,
-	CIRCLE_SELECTED,
-	CURVE_SELECTED,
-	SQUARE_SELECTED,
-	TRIANGLE_SELECTED,
-	ELIPSE_SELECTED
-};
-
-Primitive_Selection m;
