+++ CrisPaint.vcxproj
@@ -198,7 +198,6 @@
     <ClInclude Include="pch.h" />
     <ClInclude Include="Resource.h" />
     <ClInclude Include="targetver.h" />
-    <ClInclude Include="CTypes.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="CCircle.cpp" />
